fix: references is not iterable / cant parse expect(findBy*)
The bug was caused by the fact that when a query is wrapped into a function
It is not referenced in the getDeclaredVariables call making the references equal to false.
It caused the references check to be false, to enter the else statement and thus trying to iterate over a false value.